user controller use HandlerExceptionController

+++ src/main/java/it/anoki/spring/service/impl/UserServiceImpl.java
@@ -1,7 +1,5 @@
 package it.anoki.spring.service.impl;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +9,7 @@ import it.anoki.spring.repository.UserRepository;
 import it.anoki.spring.service.ReservationService;
 import it.anoki.spring.service.UserService;
 import it.anoki.spring.util.JwtTokenUtil;
+import javassist.NotFoundException;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -23,33 +22,35 @@ public class UserServiceImpl implements UserService {
 	private ReservationService reservationService;
 
 	@Override
-	public Optional<User> get(Long id) throws Exception {
-		return userRepository.findById(id);
+	public User get(Long id) throws Exception {
+		if (!userRepository.existsById(id))
+			throw new NotFoundException("User not found");
+		return userRepository.getOne(id);
 	}
 
 	@Override
 	public User save(User c) throws Exception {
+		if (c == null)
+			throw new NotFoundException("User not saved!!!");
 		return userRepository.save(c);
 	}
 
 	@Override
 	public void delete(Long id) throws Exception {
+		if (!userRepository.existsById(id))
+			throw new NotFoundException("User not found");
 		userRepository.deleteById(id);
 	}
 
 	@Override
 	public User update(Long id, String address, String email, String name) throws Exception {
-		Optional<User> u = this.get(id);
-		User user = null;
-		if (this.get(id).isPresent()) {
-			user = u.get();
-			if (address != null)
-				user.setAddress(address);
-			if (email != null)
-				user.setEmail(email);
-			if (name != null)
-				user.setName(name);
-		}
+		User user = this.get(id);
+		if (address != null)
+			user.setAddress(address);
+		if (email != null)
+			user.setEmail(email);
+		if (name != null)
+			user.setName(name);
 		return userRepository.save(user);
 	}
